networkd: fix size of MTUBytes so that it does not overwrites ARP
config_parse_iec_size overwrites the next varible that is ARP.
Now the mtu is unsigned . Make it size_t .

Fixes #4644

+++ src/network/networkd-network.h
@@ -174,7 +174,7 @@ struct Network {
         IPv6PrivacyExtensions ipv6_privacy_extensions;
 
         struct ether_addr *mac;
-        unsigned mtu;
+        size_t mtu;
         int arp;
         uint32_t iaid;
         DUID duid;
